UOP Cleanup
Some cleanup to make the unofficial patches load in the correct order.
Setting to load UOP after Shivering Isles not needed; loose files from
UOP will overwrite Russian GOTY version of the esp. Knights made to load
last due to CTDs in some setups if it is not the last official plugin.