Maskar's Oblivion Overhaul - Add to Late Changes Group
This will place it after all Bashed patches, not just the default filename.

Quote from modpage description.
This mod does not alter any game data (leveled lists, creatures, etc.) directly,
but only makes changes through OBSE scripts.
Because of this it should be fully compatible with most mods.
Load order is not important.

Note:
If this mod is installed with Wrye Bash,
this mod needs to be loaded AFTER the bashed patch.