Add mods to inc list
From Readme
The Midland Village mod placed a village on the exact same
location as Vergayun. They are incompatible with each other
and no patch will be created to resolve this.

---

A patch is available here
https://www.nexusmods.com/oblivion/mods/42864

Inc will be hidden if the patch is installed.
Added after to keep recommended order.