Update Beautiful Women
I can't see any reason for the Deactivate/NoMerge tag combo it had.
It's been on there since the BOSS masterlist import, but merging it
works just fine. Also replaced the deprecated NpcFaces tag with
NPC.FaceGen + NPC.Hair. Note that it also changes ACBS and Stats here
and there, but we don't want to forward those because it's lacking a
bunch of UOP fixes and so would override them.

Also adds a message to avoid using multiple versions of the vampires
addons.