Enforce the vanilla order for Remastered DLC plugins
The game won't load when some mod plugins are active and the DLC plugins aren't in their vanilla load order, which is different to the order used for the original Oblivion.

Since the remaster comes with all DLC this could just have a single entry for each plugin, but I decided to list them all so that the order is explicit for each plugin in isolation.